moved create match code to its own UT file
