+++ test/ApiSimTest.js
@@ -32,13 +32,7 @@ const sellParams = {
   size: 0.1,
   product_id: "LTC-USD",
 };
-const matchTemplate = {
-  side: "buy",
-  price: 32,
-  size: 2,
-  time: new Date("Dec 1 2018").toISOString(),
-  product_id: "LTC-USD",
-};
+
 describe("#ApiSim", () => {
   describe("#init", () => {
     it("has a user account sim", () => {
@@ -332,84 +326,6 @@ describe("#ApiSim", () => {
         assert.strictEqual(Gdax.currentPrice, parseFloat(data.price));
       });
       Gdax.backtest(twoCandleArray);
-    });
-  });
-
-  describe("#createMatchesFromCandle", () => {
-    it("returns a minimum of 4 matches", () => {
-      let Gdax = new ApiSim();
-      let matches = Gdax.createMatchesFromCandle(testCandleLowToHigh);
-      assert.strictEqual(matches.length, 4);
-    });
-    it("returns high before low if the close <= open", () => {
-      let Gdax = new ApiSim();
-      let matches = Gdax.createMatchesFromCandle(testCandleHighToLow);
-      assert(parseFloat(matches[1].price) >= parseFloat(matches[2].price));
-    });
-    it("returns low before high if the close > open", () => {
-      let Gdax = new ApiSim();
-      let matches = Gdax.createMatchesFromCandle(testCandleLowToHigh);
-      assert(parseFloat(matches[1].price) <= parseFloat(matches[2].price));
-    });
-    it("increases the seconds from each match by 14", () => {
-      let Gdax = new ApiSim();
-      let matches = Gdax.createMatchesFromCandle(testCandleLowToHigh);
-      for (let i = 1; i < matches.length; i++) {
-        let date = new Date(matches[i].time);
-        let datePrime = new Date(matches[i - 1].time);
-        assert.strictEqual(date.getSeconds() - datePrime.getSeconds(), 14);
-      }
-    });
-    it("changes the side to 'sell' if the price is going down from the last match", () => {
-      let Gdax = new ApiSim();
-      let matches = Gdax.createMatchesFromCandle(testCandleHighToLow);
-      assert.strictEqual(matches[2].side, "sell");
-    });
-    it("changes the side to 'buy' if the price is going up from the last match", () => {
-      let Gdax = new ApiSim();
-      let matches = Gdax.createMatchesFromCandle(testCandleHighToLow);
-      assert.strictEqual(matches[1].side, "buy");
-      assert.strictEqual(matches[3].side, "buy");
-    });
-    it("can take an array of candles and generate matches based off of them", () => {
-      let Gdax = new ApiSim();
-      let matches = Gdax.createMatchesFromCandle(twoCandleArray);
-      assert.strictEqual(matches.length, 8);
-    });
-  });
-
-  describe("#createMatch", () => {
-    it("returns a match with the specafied paramaeters", () => {
-      let Gdax = new ApiSim();
-      let match = Gdax.createMatch(matchTemplate);
-      assert.strictEqual(match.side, matchTemplate.side);
-      assert.strictEqual(match.size, matchTemplate.size);
-      assert.strictEqual(match.price, matchTemplate.price.toString());
-      assert.strictEqual(match.time, matchTemplate.time);
-      assert.strictEqual(match.product_id, matchTemplate.product_id);
-    });
-    it("returns a that generates unspecafied parameters", () => {
-      let Gdax = new ApiSim();
-      let match = Gdax.createMatch(matchTemplate);
-      assert(match.type === "match");
-      assert(typeof match.trade_id === "number");
-      assert(typeof match.sequence === "number");
-      assert(typeof match.maker_order_id === "string");
-      assert(typeof match.taker_order_id === "string");
-    });
-    it("returns the desired maker order id if specafied", () => {
-      let Gdax = new ApiSim();
-      let x = JSON.parse(JSON.stringify(matchTemplate));
-      x.maker_order_id = "abc123";
-      let match = Gdax.createMatch(x);
-      assert.strictEqual(match.maker_order_id, "abc123");
-    });
-    it("returns the desired taker order id if specafied", () => {
-      let Gdax = new ApiSim();
-      let x = JSON.parse(JSON.stringify(matchTemplate));
-      x.taker_order_id = "abc123";
-      let match = Gdax.createMatch(x);
-      assert.strictEqual(match.taker_order_id, "abc123");
     });
   });
